data_modeling_apache_cassandra/requirements.txt
numpy==1.16.3
pandas==0.24.2
pkg-resources==0.0.0
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
cassandra-driver==3.17.0
numpy==1.16.3
pandas==0.24.2
pkg-resources==0.0.0
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
